Delete Done

+++ cpp/ConfigFile.cpp
@@ -91,7 +91,7 @@ void ConfigFile::setIndex(const string& idx)
     index = idx;
 }
 
-int ConfigFile::getMax_size() const
+size_t ConfigFile::getMax_size() const
 {
     return max_size;
 }
@@ -111,7 +111,7 @@ int ConfigFile::setMax_size(const string& size)
         if (!isdigit(s[i]))
             return 0;
     }
-    int sz;
+    size_t sz;
     stringstream ss(s);
     ss >> sz;
     if (unit == 'K')
